Integer in count method explanation in CriteriaBuilder javadoc, PredicateFactory moved to sessions package + change imports
